Update dev app to allow client termination
